backend/env-mymusic/requirements.txt
Django==4.1.2
django-cors-headers==3.13.0
djangorestframework==3.14.0
numpy==1.23.4
psycopg2==2.9.4
psycopg2-binary==2.9.4
python-environ==0.4.54
python-slugify==6.1.2
pytube==12.1.0